+++ src/90_setup.h
@@ -5,6 +5,11 @@
 void setup() {
   Serial.begin(115200);
 
+  setupPowerSwitchSense();
+  handlePowerSwitchAtBoot();
+  Serial.printf("OLEG 4 Sister: SW1 sense on GPIO%d, level=%d\n",
+                SW1_SENSE_PIN, digitalRead(SW1_SENSE_PIN));
+
   analogReadResolution(12);
   analogSetPinAttenuation(BATTERY_ADC_PIN, ADC_11db);
   analogSetPinAttenuation(USB_VBUS_ADC_PIN, ADC_11db);